Drop autodetection, we don't support multiple types yet anyway
Only po files and specifically requiring use of translate.storage.pypo (and
not the other two implementations of po) is supported so far, so don't load
the files using translate.storage.factory.

The problem is that git external merge gets temporary files that don't have
the original extension, so the factory fails. When it will support multiple
classes, we'll need an option to specify the format explicitly anyway.